Fucking sqlalchemy

+++ utils.py
@@ -1,12 +1,9 @@
-from decimal import Decimal
-
-
 def num_strip(val):
     return str(val).strip().replace(" ", "").replace(",", ".")
 
 
 def parse_amount(val):
-    return Decimal(num_strip(val))
+    return float(num_strip(val))
 
 
 def parse_int(val):
